--- src/main/java/com/genius/gitget/store/item/service/ItemProvider.java
@@ -1,32 +1,0 @@
-package com.genius.gitget.store.item.service;
-
-import com.genius.gitget.global.util.exception.BusinessException;
-import com.genius.gitget.global.util.exception.ErrorCode;
-import com.genius.gitget.store.item.domain.Item;
-import com.genius.gitget.store.item.domain.ItemCategory;
-import com.genius.gitget.store.item.repository.ItemRepository;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class ItemProvider {
-    private final ItemRepository itemRepository;
-
-    public Item findById(Long itemId) {
-        return itemRepository.findById(itemId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.ITEM_NOT_FOUND));
-    }
-
-    public Item findByIdentifier(int identifier) {
-        return itemRepository.findByIdentifier(identifier)
-                .orElseThrow(() -> new BusinessException(ErrorCode.ITEM_NOT_FOUND));
-    }
-
-    public List<Item> findAllByCategory(ItemCategory itemCategory) {
-        return itemRepository.findAllByCategory(itemCategory);
-    }
-}
