+++ src/test/java/com/genius/gitget/payment/service/PaymentServiceTest.java
@@ -4,7 +4,6 @@ import static com.genius.gitget.store.item.domain.ItemCategory.CERTIFICATION_PAS
 import static com.genius.gitget.store.item.domain.ItemCategory.POINT_MULTIPLIER;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.genius.gitget.topic.repository.TopicRepository;
 import com.genius.gitget.challenge.instance.repository.InstanceRepository;
 import com.genius.gitget.challenge.likes.repository.LikesRepository;
 import com.genius.gitget.challenge.likes.service.LikesService;
@@ -20,13 +19,14 @@ import com.genius.gitget.store.item.domain.Orders;
 import com.genius.gitget.store.item.dto.ItemResponse;
 import com.genius.gitget.store.item.repository.ItemRepository;
 import com.genius.gitget.store.item.repository.OrdersRepository;
-import com.genius.gitget.store.item.service.ItemService;
+import com.genius.gitget.store.item.service.StoreFacadeImpl;
 import com.genius.gitget.store.payment.domain.Payment;
 import com.genius.gitget.store.payment.dto.PaymentDetailsResponse;
 import com.genius.gitget.store.payment.dto.PaymentRequest;
 import com.genius.gitget.store.payment.dto.PaymentResponse;
 import com.genius.gitget.store.payment.repository.PaymentRepository;
 import com.genius.gitget.store.payment.service.PaymentService;
+import com.genius.gitget.topic.repository.TopicRepository;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Assertions;
@@ -63,7 +63,7 @@ public class PaymentServiceTest {
     @Autowired
     private PaymentRepository paymentRepository;
     @Autowired
-    private ItemService itemService;
+    private StoreFacadeImpl storeFacadeImpl;
     @Autowired
     private ItemRepository itemRepository;
     @Autowired
@@ -89,7 +89,7 @@ public class PaymentServiceTest {
         getSavedOrder(user, item, itemCategory, 0);
         user.updatePoints(1000L);
 
-        ItemResponse itemResponse = itemService.orderItem(user, item.getId());
+        ItemResponse itemResponse = storeFacadeImpl.orderItem(user, item.getId());
         assertThat(itemResponse.getItemCategory()).isEqualTo(itemCategory);
 
         Page<PaymentDetailsResponse> paymentDetails = paymentService.getPaymentDetails(user, PageRequest.of(0, 10));
